% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Factorial.R
\name{Factorial}
\alias{Factorial}
\alias{A000142}
\title{Factorial numbers}
\usage{
Factorial(n, gmp = TRUE)
}
\arguments{
\item{n}{the number of first \code{n} entries from the sequence.}

\item{gmp}{a logical; \code{TRUE} to use large number representation, \code{FALSE} otherwise.}
}
\value{
a vector of length \code{n} containing first entries from the sequence.
}
\description{
Under OEIS \href{https://oeis.org/A000142}{A000142}, a \emph{Factorial} is the product of all positive integers smaller than or equal to the number.
First 6 such numbers are   	1, 1, 2, 6, 24, 120
}
\examples{
## generate first 10 Factorials
print(Factorial(10))

}
