% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Triangular.R
\name{Triangular}
\alias{Triangular}
\alias{A000217}
\title{Triangular numbers}
\usage{
Triangular(n, gmp = TRUE)
}
\arguments{
\item{n}{the number of first \code{n} entries from the sequence.}

\item{gmp}{a logical; \code{TRUE} to use large number representation, \code{FALSE} otherwise.}
}
\value{
a vector of length \code{n} containing first entries from the sequence.
}
\description{
Under OEIS \href{https://oeis.org/A000217}{A000217}, a \emph{Triangular} number counts objects arranged in an equilateral triangle.
First 6 Triangular numbers are 0, 1, 3, 6, 10, 15.
}
\examples{
## generate first 20 Triangular numbers
print(Triangular(20))

}
