% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Composite.R
\name{Composite}
\alias{Composite}
\alias{A002808}
\title{Composite numbers}
\usage{
Composite(n, gmp = TRUE)
}
\arguments{
\item{n}{the number of first \code{n} entries from the sequence.}

\item{gmp}{a logical; \code{TRUE} to use large number representation, \code{FALSE} otherwise.}
}
\value{
a vector of length \code{n} containing first entries from the sequence.
}
\description{
Under OEIS \href{https://oeis.org/A005100}{A002808}, a \emph{composite} number is
a positive integer that can be represented as multiplication of two smaller positive integers.
The first 6 composite numbers are 4, 6, 8, 9, 10, 12.
}
\examples{
## generate first 30 Composite numbers
print(Composite(30))

}
