% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unusual.R
\name{Unusual}
\alias{Unusual}
\alias{A064052}
\title{Unusual numbers}
\usage{
Unusual(n, gmp = TRUE)
}
\arguments{
\item{n}{the number of first \code{n} entries from the sequence.}

\item{gmp}{a logical; \code{TRUE} to use large number representation, \code{FALSE} otherwise.}
}
\value{
a vector of length \code{n} containing first entries from the sequence.
}
\description{
Under OEIS \href{https://oeis.org/A064052}{A064052}, an \emph{Unusual} number is a natural number whose largest prime factor is strictly greater than square root of the number.
First 6 Unusual numbers are 2, 3, 5, 6, 7, 10.
}
\examples{
## generate first 20 Unusual numbers
print(Unusual(20))

}
