% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Odious.R
\name{Odious}
\alias{Odious}
\alias{A000069}
\title{Odious numbers}
\usage{
Odious(n, Rmpfr = TRUE, PrecisionBits = 496)
}
\arguments{
\item{n}{the number of first \code{n} entries from the sequence.}

\item{Rmpfr}{a logical; \code{TRUE} to use large number representation, \code{FALSE} otherwise.}

\item{PrecisionBits}{a positive integer for precision bits larger than 2.}
}
\value{
a vector of length \code{n} containing first entries from the sequence.
}
\description{
Under OEIS \href{https://oeis.org/A000069}{A000069}, an \emph{Odious} number has an odd number of 1's in its binary expansion.
First 6 Odious numbers are   	1, 2, 4, 7, 8, 11.
}
\examples{
## generate first 20 Odious numbers
first20 = Odious(20)

## print without trailing 0's.
print(first20, drop0trailing = TRUE)

}
\seealso{
\code{\link{Evil}}
}
