% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Perfect.R
\name{Perfect}
\alias{Perfect}
\alias{A000396}
\title{Perfect numbers}
\usage{
Perfect(n, Rmpfr = TRUE, PrecisionBits = 496)
}
\arguments{
\item{n}{the number of first \code{n} entries from the sequence.}

\item{Rmpfr}{a logical; \code{TRUE} to use large number representation, \code{FALSE} otherwise.}

\item{PrecisionBits}{a positive integer for precision bits larger than 2.}
}
\value{
a vector of length \code{n} containing first entries from the sequence.
}
\description{
Under OEIS \href{https://oeis.org/A000396}{A000396}, a \emph{Perfect} number is
a number whose proper divisors sum up to the extent equal to the number itself. First
6 abundant numbers are	6, 28, 496, 8128, 33550336, 8589869056.
}
\examples{
## generate first 7 Perfect numbers
first7 = Perfect(10)

## print without trailing 0's.
print(first7, drop0trailing = TRUE)

}
\seealso{
\code{\link{Deficient}}, \code{\link{Abundant}}
}
