% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evil.R
\name{Evil}
\alias{Evil}
\alias{A001969}
\title{Evil numbers}
\usage{
Evil(n, gmp = TRUE)
}
\arguments{
\item{n}{the number of first \code{n} entries from the sequence.}

\item{gmp}{a logical; \code{TRUE} to use large number representation, \code{FALSE} otherwise.}
}
\value{
a vector of length \code{n} containing first entries from the sequence.
}
\description{
Under OEIS \href{https://oeis.org/A001969}{A001969}, an \emph{Evil} number has an even number of 1's in its binary expansion.
First 6 Evil numbers are  	0, 3, 5, 6, 9, 10.
}
\examples{
## generate first 20 Evil numbers
print(Evil(20))

}
\seealso{
\code{\link{Odious}}
}
